pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "diagram-picture-renamer"
version = "0.1.0"
description = "Strumento CLI che analizza immagini di diagrammi e genera nomi descrittivi"
authors = [ { name = "Maurizio" } ]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

dependencies = [
  "Pillow>=9.0.0",
  "typer[all]>=0.9.0",
  "python-dotenv>=1.0.0",
  "pytesseract>=0.3.10",
  "openai>=0.27.0",
  "pytest>=7.0.0",
]
